requirements.txt
flask
pvlib
pandas
requests
python-dotenv
